P4 completo funcional

+++ demo/src/main/java/com/example/demo/Entities/CreditEntity.java
@@ -22,6 +22,6 @@ public class CreditEntity {
     private double interestRate; //years
     private double monthlyPayment;  // Cuota mensual calculada
     private int state; // "Simulacion", "Solicitado", "Aprobado", "Rechazado", "Activo"
-    private int idClient;
+    private Long idClient;
 
 }
